+++ examples/shared_samples/volume_template.rb
@@ -15,7 +15,7 @@ require_relative '../_client'
 #   @storage_system_ip
 
 # All supported APIs for Volume Template:
-# - 200, 300, 500
+# - 200, 300, 500. 600
 
 raise 'Must set @storage_system_ip in _client.rb' unless @storage_system_ip
 
@@ -25,6 +25,8 @@ raise 'Must set @storage_system_ip in _client.rb' unless @storage_system_ip
 # volume_template_class = OneviewSDK::API300::Synergy::VolumeTemplate
 # volume_template_class = OneviewSDK::API500::C7000::VolumeTemplate
 # volume_template_class = OneviewSDK::API500::Synergy::VolumeTemplate
+# volume_template_class = OneviewSDK::API600::C7000::VolumeTemplate
+# volume_template_class = OneviewSDK::API600::Synergy::VolumeTemplate
 
 # Resources classes that you can use for Storage System in this example:
 # storage_system_class = OneviewSDK::API200::StorageSystem
@@ -32,6 +34,8 @@ raise 'Must set @storage_system_ip in _client.rb' unless @storage_system_ip
 # storage_system_class = OneviewSDK::API300::Synergy::StorageSystem
 # storage_system_class = OneviewSDK::API500::C7000::StorageSystem
 # storage_system_class = OneviewSDK::API500::Synergy::StorageSystem
+# storage_system_class = OneviewSDK::API600::C7000::StorageSystem
+# storage_system_class = OneviewSDK::API600::Synergy::StorageSystem
 
 # Resources classes that you can use for Storage Pool in this example:
 # storage_pool_class = OneviewSDK::API200::StoragePool
@@ -39,15 +43,23 @@ raise 'Must set @storage_system_ip in _client.rb' unless @storage_system_ip
 # storage_pool_class = OneviewSDK::API300::Synergy::StoragePool
 # storage_pool_class = OneviewSDK::API500::C7000::StoragePool
 # storage_pool_class = OneviewSDK::API500::Synergy::StoragePool
+# storage_pool_class = OneviewSDK::API600::C7000::StoragePool
+# storage_pool_class = OneviewSDK::API600::Synergy::StoragePool
 
 # Resource classses used in this sample
 volume_template_class = OneviewSDK.resource_named('VolumeTemplate', @client.api_version)
 storage_system_class = OneviewSDK.resource_named('StorageSystem', @client.api_version)
 storage_pool_class = OneviewSDK.resource_named('StoragePool', @client.api_version)
+scope_class = OneviewSDK.resource_named('Scope', @client.api_version)
+fc_network_class = OneviewSDK.resource_named('FCNetwork', @client.api_version)
+
+scope = scope_class.get_all(@client).first
+fc_network = fc_network_class.get_all(@client).first
 
 options = {
   name: 'ONEVIEW_SDK_TEST VT1',
-  description: 'Volume Template'
+  description: 'Volume Template',
+  initialScopeUris: [scope['uri']]
 }
 
 puts "\nCreating a Storage Volume Template"
@@ -85,38 +97,60 @@ else
   puts "\nListing reachable volume templates:"
   volume_template_class.get_reachable_volume_templates(@client).each { |item| puts item['name'] }
 
-  puts "\nChanging the 'isShareable' property value"
-  puts 'Before:'
-  puts volume_template['properties']['isShareable']
+  query_networks = {
+    networks: fc_network['uri']
+  }
+  puts "\nListing reachable volume templates with networks '#{query_networks[:networks]}'"
+  volume_template_class.get_reachable_volume_templates(@client, {}, query_networks).each { |item| puts item['name'] }
 
-  new_value = !volume_template.get_default_value('isShareable')
-  volume_template.set_default_value('isShareable', new_value)
-  volume_template.update
-  volume_template.refresh
+  if @client.api_version >= 600
+    query_scopes = {
+      scopeUris: scope['uri']
+    }
+    puts "\nListing reachable volume templates with scope '#{query_scopes[:scopeUris]}'"
+    volume_template_class.get_reachable_volume_templates(@client, {}, query_scopes).each { |item| puts item['name'] }
+
+    query_private_volumes = {
+      privateAllowedOnly: true
+    }
+    puts "\nListing reachable volume templates with only private volumes '#{query_private_volumes[:privateAllowedOnly]}'"
+    volume_template_class.get_reachable_volume_templates(@client, {}, query_private_volumes).each { |item| puts item['name'] }
+  end
+  volume_template_2 = volume_template_class.new(@client, name: options[:name])
+  volume_template_2.retrieve!
+  puts "\nChanging the 'isShareaitem2ble' property value"
+  puts 'Before:'
+  puts volume_template_2['properties']['isShareable']
+
+  new_value = !volume_template_2.get_default_value('isShareable')
+  volume_template_2.set_default_value('isShareable', new_value)
+  volume_template_2.update
+  volume_template_2.retrieve!
   puts 'After:'
-  puts volume_template['properties']['isShareable']
+  puts volume_template_2['properties']['isShareable']
 
   puts "\nChanging the locked property of 'isShareable'"
   puts 'Before:'
-  puts volume_template['properties']['isShareable']['meta']
+  puts volume_template_2['properties']['isShareable']['meta']
 
-  volume_template.lock('isShareable')
-  volume_template.update
-  volume_template.refresh
+  volume_template_2.lock('isShareable')
+  volume_template_2.update
+  volume_template_2.retrieve!
   puts 'After:'
-  puts volume_template['properties']['isShareable']['meta']
+  puts volume_template_2['properties']['isShareable']['meta']
 end
 
 # Retrieve created volume template
-volume_template_2 = volume_template_class.new(@client, name: options[:name])
-volume_template_2.retrieve!
-puts "\nRetrieved Volume Template by name: '#{volume_template_2[:name]}'.\n  uri = '#{volume_template_2[:uri]}'"
+volume_template_3 = volume_template_class.new(@client, name: options[:name])
+volume_template_3.retrieve!
+puts "\nRetrieved Volume Template by name: '#{volume_template_3[:name]}'.\n  uri = '#{volume_template_3[:uri]}'"
 
 # Find recently created volume template by name
 matches = volume_template_class.find_by(@client, name: options[:name])
-volume_template_3 = matches.first
-puts "\nFound Volume Template by name: '#{volume_template_3[:name]}'.\n  uri = '#{volume_template_3[:uri]}'"
+volume_template_4 = matches.first
+puts "\nFound Volume Template by name: '#{volume_template_4[:name]}'.\n  uri = '#{volume_template_4[:uri]}'"
 
 # Delete Volume Template
+volume_template.retrieve!
 volume_template.delete
 puts "\nDeleted Volume Template '#{volume_template[:name]}' successfully.\n"