+++ spec/integration/resource/api600/c7000/volume_template/create_spec.rb
@@ -11,11 +11,11 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::VolumeTemplate
+klass = OneviewSDK::API600::C7000::VolumeTemplate
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  let(:current_client) { $client_500 }
+  let(:current_client) { $client_600 }
   let(:storage_system_ip) { $secrets['storage_system1_ip'] }
   let(:storage_virtual_ip) { $secrets['store_virtual_ip'] }
-  include_context 'integration api500 context'
-  include_examples 'VolumeTemplateCreateExample API500'
+  include_context 'integration api600 context'
+  include_examples 'VolumeTemplateCreateExample API600'
 end